+++ Systems/Engine/DataSource.hpp
@@ -169,18 +169,18 @@ class ArrayDataSource : public DataSource
 {
 public:
 
-  static const u64 RootIndex = (u64)-1;
+  static const size_t RootIndex = (size_t)-1;
 
   /// Return the size of the contained array
-  virtual u32 GetArraySize() = 0;
+  virtual size_t GetArraySize() = 0;
   /// Fills out the variant with the data of the given column at the given array index
-  virtual void GetData(u32 index,       Any& any, StringParam column) = 0;
+  virtual void GetData(size_t index,       Any& any, StringParam column) = 0;
   /// Sets the data of the given column at the given array index with the passed in variant.
-  virtual void SetData(u32 index, const Any& any, StringParam column) = 0;
+  virtual void SetData(size_t index, const Any& any, StringParam column) = 0;
 
   // Helpers to get an array index from an entry or index
-  u32 DataEntryToArrayIndex(DataEntry* entry);
-  u32 DataIndexToArrayIndex(DataIndex index);
+  size_t DataEntryToArrayIndex(DataEntry* entry);
+  size_t DataIndexToArrayIndex(DataIndex index);
 
   // DataSource Interface
   DataEntry* GetRoot() override;
@@ -269,7 +269,7 @@ public:
 
   uint Size() override
   {
-    return mSelection.Size();
+    return (uint)mSelection.Size();
   }
 };
 
@@ -365,7 +365,7 @@ public:
 
   void SetSource(containerType* data){mData = data;}
   virtual bool IsCollection(){return true;}
-  virtual uint GetCount(){return mData->Size();}
+  virtual uint GetCount(){return (uint)mData->Size();}
 
   uint Size(){return mData->Size();}
   type& operator[](uint index){return (*mData)[index];}
