+++ Systems/Engine/Spline.cpp
@@ -161,7 +161,7 @@ void SplineControlPoints::Clear()
 
 int SplineControlPoints::GetCount() const
 {
-  return mControlPoints.Size();
+  return (int)mControlPoints.Size();
 }
 
 //-------------------------------------------------------------------SplineBakedPoint
@@ -393,7 +393,7 @@ void Spline::DebugDraw(Vec4Param color)
 {
   RebuildIfModified();
 
-  for(size_t i = 1; i < mBakedCurve.Size(); ++i)
+  for(uint i = 1; i < (uint)mBakedCurve.Size(); ++i)
   {
     Vec3 p0 = mBakedCurve.GetPoint(i - 1).Position;
     Vec3 p1 = mBakedCurve.GetPoint(i).Position;
